requirements_test.txt
coverage==4.5.3
pytest==6.1.0
pytest-django==3.10.0
pytest-cov==2.7.1
freezegun==0.3.12
pytest-sugar==0.9.4
factory-boy==3.0.1
pytest-mock==3.2.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
